Refactor tests into separate modules for improved organization
- Moved tests from `core/engine/src/module/loader/mod.rs` to `core/engine/src/module/loader/tests.rs`.
- Moved tests from `core/engine/src/string.rs` to `core/engine/src/string/tests.rs`.
- Moved tests from `core/engine/src/value/conversions/serde_json.rs` to `core/engine/src/value/conversions/tests.rs`.
- Moved tests from `core/engine/src/value/integer.rs` to `core/engine/src/value/integer/tests.rs`.
- Moved tests from `core/macros/src/utils.rs` to `core/macros/src/utils/tests.rs`.

This restructuring enhances code readability and maintainability by isolating test cases in dedicated files.
