+++ core/engine/src/builtins/element.rs
@@ -1289,415 +1289,4 @@ fn attach_shadow(this: &JsValue, args: &[JsValue], context: &mut Context) -> JsR
 }
 
 #[cfg(test)]
-mod tests {
-    use crate::{js_string, run_test_actions, TestAction, JsNativeErrorKind, JsValue};
-
-    #[test]
-    fn element_constructor() {
-        run_test_actions([
-            TestAction::assert_eq("typeof Element", js_string!("function")),
-            TestAction::run("var elem = new Element()"),
-            TestAction::assert_eq("typeof elem", js_string!("object")),
-            TestAction::assert_eq("elem !== null", true),
-        ]);
-    }
-
-    #[test]
-    fn element_constructor_allows_direct_call() {
-        run_test_actions([
-            // This implementation allows direct call (not WHATWG compliant but existing behavior)
-            TestAction::run("var elem = Element()"),
-            TestAction::assert_eq("typeof elem", js_string!("object")),
-            TestAction::assert_eq("elem !== null", true),
-        ]);
-    }
-
-    #[test]
-    fn element_inheritance() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Test constructor relationship
-            TestAction::assert_eq("elem.constructor === Element", true),
-            TestAction::assert_eq("elem instanceof Element", true),
-
-            // Test that it's an object
-            TestAction::assert_eq("typeof elem", js_string!("object")),
-            TestAction::assert_eq("elem !== null", true),
-        ]);
-    }
-
-    #[test]
-    fn element_methods_exist() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Test all required methods exist
-            TestAction::assert_eq("typeof elem.getAttribute", js_string!("function")),
-            TestAction::assert_eq("typeof elem.setAttribute", js_string!("function")),
-            TestAction::assert_eq("typeof elem.removeAttribute", js_string!("function")),
-            TestAction::assert_eq("typeof elem.hasAttribute", js_string!("function")),
-            TestAction::assert_eq("typeof elem.appendChild", js_string!("function")),
-            TestAction::assert_eq("typeof elem.removeChild", js_string!("function")),
-            TestAction::assert_eq("typeof elem.setHTML", js_string!("function")),
-            TestAction::assert_eq("typeof elem.setHTMLUnsafe", js_string!("function")),
-            TestAction::assert_eq("typeof elem.attachShadow", js_string!("function")),
-        ]);
-    }
-
-    #[test]
-    fn element_properties_exist() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Test properties (these are getters/setters)
-            TestAction::assert_eq("'tagName' in elem", true),
-            TestAction::assert_eq("'id' in elem", true),
-            TestAction::assert_eq("'className' in elem", true),
-            TestAction::assert_eq("'innerHTML' in elem", true),
-            TestAction::assert_eq("'textContent' in elem", true),
-            TestAction::assert_eq("'children' in elem", true),
-            TestAction::assert_eq("'parentNode' in elem", true),
-            TestAction::assert_eq("'style' in elem", true),
-        ]);
-    }
-
-    #[test]
-    fn element_tag_name() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Default element has empty tag name in this implementation
-            TestAction::assert_eq("typeof elem.tagName", js_string!("string")),
-            TestAction::assert_eq("elem.tagName", js_string!("")),
-
-            // tagName should be read-only (setter doesn't exist)
-            TestAction::run("try { elem.tagName = 'DIV'; } catch(e) {}"),
-            TestAction::assert_eq("elem.tagName", js_string!("")),
-        ]);
-    }
-
-    #[test]
-    fn element_id_property() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Initially empty
-            TestAction::assert_eq("elem.id", js_string!("")),
-
-            // Set and get
-            TestAction::run("elem.id = 'test-id'"),
-            TestAction::assert_eq("elem.id", js_string!("test-id")),
-
-            // Clear
-            TestAction::run("elem.id = ''"),
-            TestAction::assert_eq("elem.id", js_string!("")),
-        ]);
-    }
-
-    #[test]
-    fn element_class_name_property() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Initially empty
-            TestAction::assert_eq("elem.className", js_string!("")),
-
-            // Set and get
-            TestAction::run("elem.className = 'class1 class2'"),
-            TestAction::assert_eq("elem.className", js_string!("class1 class2")),
-
-            // Clear
-            TestAction::run("elem.className = ''"),
-            TestAction::assert_eq("elem.className", js_string!("")),
-        ]);
-    }
-
-    #[test]
-    fn element_attribute_methods() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Initially no attributes
-            TestAction::assert_eq("elem.hasAttribute('data-test')", false),
-            TestAction::assert_eq("elem.getAttribute('data-test')", JsValue::null()),
-
-            // Set attribute
-            TestAction::run("elem.setAttribute('data-test', 'value1')"),
-            TestAction::assert_eq("elem.hasAttribute('data-test')", true),
-            TestAction::assert_eq("elem.getAttribute('data-test')", js_string!("value1")),
-
-            // Update attribute
-            TestAction::run("elem.setAttribute('data-test', 'value2')"),
-            TestAction::assert_eq("elem.getAttribute('data-test')", js_string!("value2")),
-
-            // Remove attribute
-            TestAction::run("elem.removeAttribute('data-test')"),
-            TestAction::assert_eq("elem.hasAttribute('data-test')", false),
-            TestAction::assert_eq("elem.getAttribute('data-test')", JsValue::null()),
-        ]);
-    }
-
-    #[test]
-    fn element_multiple_attributes() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Set multiple attributes
-            TestAction::run("elem.setAttribute('id', 'test-id')"),
-            TestAction::run("elem.setAttribute('class', 'test-class')"),
-            TestAction::run("elem.setAttribute('data-value', '123')"),
-
-            // All should exist
-            TestAction::assert_eq("elem.hasAttribute('id')", true),
-            TestAction::assert_eq("elem.hasAttribute('class')", true),
-            TestAction::assert_eq("elem.hasAttribute('data-value')", true),
-
-            // Values should be correct
-            TestAction::assert_eq("elem.getAttribute('id')", js_string!("test-id")),
-            TestAction::assert_eq("elem.getAttribute('class')", js_string!("test-class")),
-            TestAction::assert_eq("elem.getAttribute('data-value')", js_string!("123")),
-
-            // Remove one
-            TestAction::run("elem.removeAttribute('class')"),
-            TestAction::assert_eq("elem.hasAttribute('id')", true),
-            TestAction::assert_eq("elem.hasAttribute('class')", false),
-            TestAction::assert_eq("elem.hasAttribute('data-value')", true),
-        ]);
-    }
-
-    #[test]
-    fn element_inner_html() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Initially empty
-            TestAction::assert_eq("elem.innerHTML", js_string!("")),
-
-            // Set HTML content
-            TestAction::run("elem.innerHTML = '<span>Hello</span>'"),
-            TestAction::assert_eq("elem.innerHTML", js_string!("<span>Hello</span>")),
-
-            // Clear content
-            TestAction::run("elem.innerHTML = ''"),
-            TestAction::assert_eq("elem.innerHTML", js_string!("")),
-        ]);
-    }
-
-    #[test]
-    fn element_text_content() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Initially empty
-            TestAction::assert_eq("elem.textContent", js_string!("")),
-
-            // Set text content
-            TestAction::run("elem.textContent = 'Hello World'"),
-            TestAction::assert_eq("elem.textContent", js_string!("Hello World")),
-
-            // Clear content
-            TestAction::run("elem.textContent = ''"),
-            TestAction::assert_eq("elem.textContent", js_string!("")),
-        ]);
-    }
-
-    #[test]
-    fn element_children_property() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Initially no children
-            TestAction::assert_eq("Array.isArray(elem.children)", true),
-            TestAction::assert_eq("elem.children.length", 0),
-        ]);
-    }
-
-    #[test]
-    fn element_parent_node() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Initially no parent
-            TestAction::assert_eq("elem.parentNode", JsValue::null()),
-        ]);
-    }
-
-    #[test]
-    fn element_style_property() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Style should be an object
-            TestAction::assert_eq("typeof elem.style", js_string!("object")),
-            TestAction::assert_eq("elem.style !== null", true),
-        ]);
-    }
-
-    #[test]
-    fn element_append_remove_child() {
-        run_test_actions([
-            TestAction::run("var parent = new Element()"),
-            TestAction::run("var child = new Element()"),
-
-            // Initially no children
-            TestAction::assert_eq("parent.children.length", 0),
-            TestAction::assert_eq("child.parentNode", JsValue::null()),
-
-            // Append child
-            TestAction::run("var result = parent.appendChild(child)"),
-            TestAction::assert_eq("result === child", true),
-            TestAction::assert_eq("parent.children.length", 1),
-            TestAction::assert_eq("parent.children[0] === child", true),
-            TestAction::assert_eq("child.parentNode === parent", true),
-
-            // Remove child
-            TestAction::run("var removed = parent.removeChild(child)"),
-            TestAction::assert_eq("removed === child", true),
-            TestAction::assert_eq("parent.children.length", 0),
-            TestAction::assert_eq("child.parentNode", JsValue::null()),
-        ]);
-    }
-
-    #[test]
-    fn element_set_html_methods() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Test setHTML method
-            TestAction::run("elem.setHTML('<p>Safe HTML</p>')"),
-            TestAction::assert_eq("elem.innerHTML", js_string!("<p>Safe HTML</p>")),
-
-            // Test setHTMLUnsafe method
-            TestAction::run("elem.setHTMLUnsafe('<script>alert(\"test\")</script>')"),
-            TestAction::assert_eq("elem.innerHTML", js_string!("<script>alert(\"test\")</script>")),
-        ]);
-    }
-
-    // Shadow DOM tests disabled due to borrowing conflicts in current implementation
-    // The attachShadow method exists but has internal borrowing issues
-
-    #[test]
-    fn element_method_errors() {
-        run_test_actions([
-            // Test methods on non-object
-            TestAction::assert_native_error(
-                "Element.prototype.getAttribute.call(null, 'test')",
-                JsNativeErrorKind::Type,
-                "Element.prototype.getAttribute called on non-object",
-            ),
-
-            TestAction::assert_native_error(
-                "Element.prototype.setAttribute.call(null, 'test', 'value')",
-                JsNativeErrorKind::Type,
-                "Element.prototype.setAttribute called on non-object",
-            ),
-
-            TestAction::assert_native_error(
-                "Element.prototype.removeAttribute.call(null, 'test')",
-                JsNativeErrorKind::Type,
-                "Element.prototype.removeAttribute called on non-object",
-            ),
-
-            TestAction::assert_native_error(
-                "Element.prototype.hasAttribute.call(null, 'test')",
-                JsNativeErrorKind::Type,
-                "Element.prototype.hasAttribute called on non-object",
-            ),
-        ]);
-    }
-
-    #[test]
-    fn element_property_errors() {
-        run_test_actions([
-            TestAction::run("var obj = {}"),
-
-            // Test property access on non-Element object
-            TestAction::assert_native_error(
-                "Object.getOwnPropertyDescriptor(Element.prototype, 'tagName').get.call(obj)",
-                JsNativeErrorKind::Type,
-                "Element.prototype.tagName called on non-Element object",
-            ),
-
-            TestAction::assert_native_error(
-                "Object.getOwnPropertyDescriptor(Element.prototype, 'id').get.call(obj)",
-                JsNativeErrorKind::Type,
-                "Element.prototype.id called on non-Element object",
-            ),
-
-            TestAction::assert_native_error(
-                "Object.getOwnPropertyDescriptor(Element.prototype, 'className').get.call(obj)",
-                JsNativeErrorKind::Type,
-                "Element.prototype.className called on non-Element object",
-            ),
-        ]);
-    }
-
-    #[test]
-    fn element_interface_compliance() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // Test all required interface members exist
-            TestAction::assert_eq("'tagName' in elem", true),
-            TestAction::assert_eq("'id' in elem", true),
-            TestAction::assert_eq("'className' in elem", true),
-            TestAction::assert_eq("'innerHTML' in elem", true),
-            TestAction::assert_eq("'textContent' in elem", true),
-            TestAction::assert_eq("'children' in elem", true),
-            TestAction::assert_eq("'parentNode' in elem", true),
-            TestAction::assert_eq("'style' in elem", true),
-            TestAction::assert_eq("'getAttribute' in elem", true),
-            TestAction::assert_eq("'setAttribute' in elem", true),
-            TestAction::assert_eq("'removeAttribute' in elem", true),
-            TestAction::assert_eq("'hasAttribute' in elem", true),
-            TestAction::assert_eq("'appendChild' in elem", true),
-            TestAction::assert_eq("'removeChild' in elem", true),
-        ]);
-    }
-
-    #[test]
-    fn element_id_attribute_sync() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // In this implementation, property and attribute are separate
-            // Setting id property doesn't automatically set attribute
-            TestAction::run("elem.id = 'test-id'"),
-            TestAction::assert_eq("elem.id", js_string!("test-id")),
-            TestAction::assert_eq("elem.getAttribute('id')", JsValue::null()), // Not synced
-
-            // Setting attribute doesn't automatically set property
-            TestAction::run("elem.setAttribute('id', 'attr-id')"),
-            TestAction::assert_eq("elem.getAttribute('id')", js_string!("attr-id")),
-            TestAction::assert_eq("elem.id", js_string!("test-id")), // Still old value
-
-            // They are independent storage mechanisms
-            TestAction::run("elem.removeAttribute('id')"),
-            TestAction::assert_eq("elem.getAttribute('id')", JsValue::null()),
-            TestAction::assert_eq("elem.id", js_string!("test-id")), // Property unchanged
-        ]);
-    }
-
-    #[test]
-    fn element_class_attribute_sync() {
-        run_test_actions([
-            TestAction::run("var elem = new Element()"),
-
-            // In this implementation, className property and class attribute are separate
-            TestAction::run("elem.className = 'class1 class2'"),
-            TestAction::assert_eq("elem.className", js_string!("class1 class2")),
-            TestAction::assert_eq("elem.getAttribute('class')", JsValue::null()), // Not synced
-
-            // Setting attribute doesn't automatically set property
-            TestAction::run("elem.setAttribute('class', 'attr-class')"),
-            TestAction::assert_eq("elem.getAttribute('class')", js_string!("attr-class")),
-            TestAction::assert_eq("elem.className", js_string!("class1 class2")), // Still old value
-
-            // They are independent storage mechanisms
-            TestAction::run("elem.removeAttribute('class')"),
-            TestAction::assert_eq("elem.getAttribute('class')", JsValue::null()),
-            TestAction::assert_eq("elem.className", js_string!("class1 class2")), // Property unchanged
-        ]);
-    }
-}
+mod tests;
